+++ common/src/View/UVView.h
@@ -112,7 +112,7 @@ namespace TrenchBroom {
             void processEvent(const MouseEvent& event) override;
             void processEvent(const CancelEvent& event) override;
         private:
-            PickRequest doGetPickRequest(int x, int y) const override;
+            PickRequest doGetPickRequest(float x, float y) const override;
             Model::PickResult doPick(const vm::ray3& pickRay) const override;
         };
     }
